@@ -1,40 +1,51 @@
 package com.sungardas.enhancedsnapshots.components;
 
+import java.util.Comparator;
+import java.util.List;
+import java.util.Set;
+import java.util.TreeSet;
+import java.util.concurrent.ExecutorService;
+import java.util.concurrent.Executors;
+import java.util.concurrent.TimeUnit;
+
+import javax.annotation.PostConstruct;
+import javax.annotation.PreDestroy;
+
 import com.amazonaws.AmazonClientException;
 import com.sungardas.enhancedsnapshots.aws.dynamodb.model.TaskEntry;
 import com.sungardas.enhancedsnapshots.aws.dynamodb.repository.TaskRepository;
 import com.sungardas.enhancedsnapshots.exception.EnhancedSnapshotsInterruptedException;
-import com.sungardas.enhancedsnapshots.service.ConfigurationService;
 import com.sungardas.enhancedsnapshots.service.NotificationService;
 import com.sungardas.enhancedsnapshots.service.SDFSStateService;
+import com.sungardas.enhancedsnapshots.service.SystemService;
 import com.sungardas.enhancedsnapshots.service.TaskService;
 import com.sungardas.enhancedsnapshots.tasks.executors.TaskExecutor;
+
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
-import org.springframework.beans.factory.annotation.*;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.beans.factory.annotation.Qualifier;
 import org.springframework.context.annotation.DependsOn;
 import org.springframework.stereotype.Service;
 
-import javax.annotation.*;
-import java.util.*;
-import java.util.concurrent.*;
-import static com.sungardas.enhancedsnapshots.aws.dynamodb.model.TaskEntry.TaskEntryStatus.*;
+import static com.sungardas.enhancedsnapshots.aws.dynamodb.model.TaskEntry.TaskEntryStatus.ERROR;
+import static com.sungardas.enhancedsnapshots.aws.dynamodb.model.TaskEntry.TaskEntryStatus.RUNNING;
+
 
 @Service
-@DependsOn("CreateAppConfiguration")
+@DependsOn("SystemService")
 public class WorkersDispatcher {
-    private static final Comparator<TaskEntry> taskComparatorByTimeAndPriority = new Comparator<TaskEntry>() {
-        @Override
-        public int compare(TaskEntry o1, TaskEntry o2) {
-            int priority = o2.getPriority() - o1.getPriority();
-            if (priority != 0) {
-                return priority;
-            }
-            return o1.getSchedulerTime().compareTo(o2.getSchedulerTime());
+
+    private static final Comparator<TaskEntry> taskComparatorByTimeAndPriority = (o1, o2) -> {
+        int priority = o2.getPriority() - o1.getPriority();
+        if (priority != 0) {
+            return priority;
         }
+        return o1.getSchedulerTime().compareTo(o2.getSchedulerTime());
     };
+
     @Autowired
-    private ConfigurationService configurationService;
+    private ConfigurationMediator configurationMediator;
     @Autowired
     @Qualifier("awsBackupVolumeTaskExecutor")
     private TaskExecutor awsBackupVolumeTaskExecutor;
@@ -48,6 +59,8 @@ public class WorkersDispatcher {
     private TaskService taskService;
     @Autowired
     private SDFSStateService sdfsStateService;
+    @Autowired
+    private SystemService systemService;
     @Autowired
     private TaskRepository taskRepository;
     @Autowired
@@ -79,8 +92,6 @@ public class WorkersDispatcher {
 
         @Override
         public void run() {
-            String instanceId = configurationService.getConfigurationId();
-
             LOGtw.info("Starting worker dispatcher");
             while (true) {
                 if (Thread.interrupted()) {
@@ -88,13 +99,17 @@ public class WorkersDispatcher {
                 }
                 TaskEntry entry = null;
                 try {
-                    Set<TaskEntry> taskEntrySet = sortByTimeAndPriority(taskRepository.findByStatusAndInstanceIdAndRegular(TaskEntry.TaskEntryStatus.QUEUED.getStatus(), instanceId, Boolean.FALSE.toString()));
+                    Set<TaskEntry> taskEntrySet = sortByTimeAndPriority(taskRepository.findByStatusAndRegular(TaskEntry.TaskEntryStatus.QUEUED.getStatus(), Boolean.FALSE.toString()));
                     while (!taskEntrySet.isEmpty()) {
                         entry = taskEntrySet.iterator().next();
 
                         if (!taskService.isCanceled(entry.getId())) {
                             switch (TaskEntry.TaskEntryType.getType(entry.getType())) {
                                 case BACKUP:
+                                    //TODO remove when sdfscli will allow to expand cache size at runtime
+                                    if (!sdfsStateService.sdfsIsAvailable()) {
+                                        break;
+                                    }
                                     LOGtw.info("Task was identified as backup");
                                     awsBackupVolumeTaskExecutor.execute(entry);
                                     break;
@@ -104,6 +119,9 @@ public class WorkersDispatcher {
                                     break;
                                 }
                                 case RESTORE:
+                                    if (!sdfsStateService.sdfsIsAvailable()) {
+                                        break;
+                                    }
                                     LOGtw.info("Task was identified as restore");
                                     awsRestoreVolumeTaskExecutor.execute(entry);
                                     break;
@@ -112,7 +130,7 @@ public class WorkersDispatcher {
                                     notificationService.notifyAboutTaskProgress(entry.getId(), "System backup started", 0);
                                     entry.setStatus(RUNNING.getStatus());
                                     taskRepository.save(entry);
-                                    sdfsStateService.backupState(entry.getId());
+                                    systemService.backup(entry.getId());
                                     taskRepository.delete(entry);
                                     notificationService.notifyAboutTaskProgress(entry.getId(), "System backup finished", 100);
                                     break;
@@ -124,7 +142,7 @@ public class WorkersDispatcher {
                         } else {
                             LOGtw.debug("Task canceled: {}", entry);
                         }
-                        taskEntrySet = sortByTimeAndPriority(taskRepository.findByStatusAndInstanceIdAndRegular(TaskEntry.TaskEntryStatus.QUEUED.getStatus(), instanceId, Boolean.FALSE.toString()));
+                        taskEntrySet = sortByTimeAndPriority(taskRepository.findByStatusAndRegular(TaskEntry.TaskEntryStatus.QUEUED.getStatus(), Boolean.FALSE.toString()));
                     }
                     sleep();
                 } catch (AmazonClientException e) {
@@ -147,7 +165,7 @@ public class WorkersDispatcher {
 
         private void sleep() {
             try {
-                TimeUnit.MILLISECONDS.sleep(configurationService.getWorkerDispatcherPollingRate());
+                TimeUnit.MILLISECONDS.sleep(configurationMediator.getWorkerDispatcherPollingRate());
             } catch (InterruptedException e) {
                 e.printStackTrace();
             }
