@@ -3,9 +3,9 @@
 angular.module('web')
     .controller('ConfigController', function ($scope, Volumes, Configuration, $modal, $state) {
         var DELAYTIME = 600*1000;
-            $scope.isBusy = false;
         $scope.STRINGS = {
             s3: {
+                empty: 'Bucket name field cannot be empty',
                 new: 'Will be created new as',
                 existing: 'Will be used existing bucket:'
             },
@@ -38,58 +38,45 @@ angular.module('web')
             false: 'danger'
         };
 
-        $scope.isAWS = false;
+        $scope.isCustomBucketName = false;
+        $scope.isNameWrong = false;
+        $scope.wrongNameMessage = '';
         $scope.isValidInstance = true;
         $scope.selectBucket = function (bucket) {
             $scope.selectedBucket = bucket;
         };
 
-        var getAwsStatus = function () {
-            $scope.isBusy = true;
-            Configuration.get('awscreds').then(function (result, status) {
-                if (result.data.contains) {
-                    getCurrentConfig();
-                }
-                else{
-                    $scope.isAWS = true;
-                    $scope.isBusy = false;
-                }
-            }, function (data, status) {
-                $scope.isValidInstance = false;
-                $scope.invalidMessage = data.data.localizedMessage;
-                $scope.isBusy = false;
+        var wizardCreationProgress = function () {
+            var modalInstance = $modal.open({
+                animation: true,
+                templateUrl: './partials/modal.wizard-progress.html',
+                scope: $scope
             });
+
+            modalInstance.result.then(function () {
+                $state.go('login')
+            }, function () {
+            });
+
+            return modalInstance
         };
-        getAwsStatus();
 
         var getCurrentConfig = function () {
+            $scope.progressState = 'loading';
+            var loader = wizardCreationProgress();
+
             Configuration.get('current').then(function (result, status) {
                 $scope.settings = result.data;
                 $scope.selectedBucket = (result.data.s3 || [])[0] || {};
-                $scope.isAWS = false;
-                $scope.isBusy = false;
+                loader.dismiss();
             }, function (data, status) {
                 $scope.isValidInstance = false;
                 $scope.invalidMessage = data.data.localizedMessage;
-                $scope.isBusy = false;
-            })
-        };
-
-        $scope.awsPublicKey = '';
-        $scope.awsSecretKey = '';
-
-        $scope.sendAWS = function () {
-            var keysCollection = {
-                'awsPublicKey': $scope.awsPublicKey,
-                'awsSecretKey': $scope.awsSecretKey
-            };
-
-            Configuration.send('awscreds', keysCollection).then(function () {
-                getCurrentConfig();
-
-            }, function (data, status) {
+                loader.dismiss();
             });
         };
+
+        getCurrentConfig();
 
         $scope.sendSettings = function () {
             var volumeSize = $scope.isNewVolumeSize ? $scope.sdfsNewSize : $scope.settings.sdfs.volumeSize;
@@ -105,13 +92,13 @@ angular.module('web')
                     admin: true
                 };
 
-                var awsModalInstance = $modal.open({
+                var userModalInstance = $modal.open({
                     animation: true,
                     templateUrl: './partials/modal.user-edit.html',
                     scope: $scope
                 });
 
-                awsModalInstance.result.then(function () {
+                userModalInstance.result.then(function () {
                     settings.user = $scope.userToEdit;
                     delete settings.user.isNew;
 
@@ -130,7 +117,7 @@ angular.module('web')
 
                 Configuration.send('current', settings).then(function () {
                     $scope.progressState = 'success';
-                }, function () {
+                }, function (data, status) {
                     $scope.progressState = 'failed';
                 });
 
@@ -138,16 +125,13 @@ angular.module('web')
             }
         };
 
-        var wizardCreationProgress = function () {
-            var modalInstance = $modal.open({
-                animation: true,
-                templateUrl: './partials/modal.wizard-progress.html',
-                scope: $scope
+        $scope.validateName = function () {
+            Configuration.get('bucket/' + encodeURIComponent($scope.selectedBucket.bucketName)).then(function (result) {
+                $scope.isNameWrong = !result.data.valid;
+                $scope.wrongNameMessage = result.data.message;
+            }, function (data, status) {
             });
-
-            modalInstance.result.then(function () {
-                $state.go('login')
-            })
         };
+
 
     });
