@@ -5,14 +5,45 @@ angular.module('web')
         var currentUser = Users.getCurrent();
         $scope.isAdmin = currentUser.role === "admin";
 
+        $scope.STRINGS = {
+            sdfs: {
+                sdfsLocalCacheSize: {
+                  empty: 'Local Cache Size field cannot be empty.'
+                } ,
+                volumeSize: {
+                 empty: 'Volume Size field cannot be empty.'
+                }
+            },
+            volumeType: {
+               empty: 'Volume size for io1 volume type cannot be empty.',
+               range: 'Volume size for io1 volume type must be in between 1 and 30.'
+            },
+            otherSettings: {
+                empty: 'All fields are required. Please fill in empty fields.'
+            }
+        };
+
+        var progressLoader = function () {
+            var modalInstance = $modal.open({
+                animation: true,
+                templateUrl: './partials/modal.wizard-progress.html',
+                scope: $scope
+            });
+
+            return modalInstance
+        };
+
+        $scope.progressState = 'loading';
+        var loader = progressLoader();
         System.get().then(function (data) {
             $scope.settings = data;
-            $scope.initialTempVolumeType = data.systemProperties.tempVolumeType;
-            $scope.initialRestoreVolumeType = data.systemProperties.restoreVolumeType;
-            $scope.initialTempVolumeSize = data.systemProperties.tempVolumeIopsPerGb;
-            $scope.initialRestoreVolumeSize = data.systemProperties.restoreVolumeIopsPerGb;
+            $scope.initialSettings = angular.copy(data);
+            $scope.progressState = '';
+            loader.dismiss();
         }, function (e) {
             console.log(e);
+            $scope.progressState = 'failed';
+            loader.dismiss();
         });
 
         $scope.backup = function () {
@@ -34,19 +65,25 @@ angular.module('web')
 
         };
 
-        $scope.changeType = function () {
-            var typeChangeConfirmationModal = $modal.open({
+        $scope.updateSettings = function () {
+
+            var settingsUpdateModal = $modal.open({
                 animation: true,
                 scope: $scope,
-                templateUrl: './partials/modal.volume-type.html',
+                templateUrl: './partials/modal.settings-update.html',
                 controller: 'modalVolumeTypeChangeCtrl'
             });
 
-            typeChangeConfirmationModal.result.then(function () {
-                $scope.initialTempVolumeType = $scope.settings.systemProperties.tempVolumeType;
-                $scope.initialRestoreVolumeType = $scope.settings.systemProperties.restoreVolumeType;
-                $scope.initialTempVolumeSize = $scope.settings.systemProperties.tempVolumeIopsPerGb;
-                $scope.initialRestoreVolumeSize = $scope.settings.systemProperties.restoreVolumeIopsPerGb;
+            settingsUpdateModal.result.then(function () {
+                if ($scope.state == "done") {
+                    $scope.initialSettings = angular.copy($scope.settings);
+                }
             });
         };
+
+        $scope.isNewValues = function () {
+           return JSON.stringify($scope.settings) !== JSON.stringify($scope.initialSettings);
+        };
+
+
     });
