@@ -5,11 +5,15 @@ import com.amazonaws.regions.Region;
 import com.amazonaws.regions.Regions;
 import com.amazonaws.services.dynamodbv2.AmazonDynamoDB;
 import com.amazonaws.services.dynamodbv2.AmazonDynamoDBClient;
+import com.amazonaws.services.dynamodbv2.datamodeling.DynamoDBMapper;
+import com.amazonaws.services.dynamodbv2.datamodeling.DynamoDBMapperConfig;
 import com.amazonaws.services.ec2.AmazonEC2;
 import com.amazonaws.services.ec2.AmazonEC2Client;
 import com.amazonaws.services.s3.AmazonS3;
 import com.amazonaws.services.s3.AmazonS3Client;
+import com.amazonaws.util.EC2MetadataUtils;
 import com.sungardas.enhancedsnapshots.components.RetryInterceptor;
+
 import org.socialsignin.spring.data.dynamodb.repository.config.EnableDynamoDBRepositories;
 import org.springframework.aop.framework.ProxyFactoryBean;
 import org.springframework.context.annotation.Bean;
@@ -18,7 +22,7 @@ import org.springframework.context.annotation.Profile;
 
 @Configuration
 @Profile("prod")
-@EnableDynamoDBRepositories(basePackages = "com.sungardas.enhancedsnapshots.aws.dynamodb.repository")
+@EnableDynamoDBRepositories(basePackages = "com.sungardas.enhancedsnapshots.aws.dynamodb.repository", dynamoDBMapperConfigRef = "dynamoDBMapperConfig")
 public class AmazonConfigProvider {
     private InstanceProfileCredentialsProvider  credentialsProvider;
 
@@ -64,11 +68,33 @@ public class AmazonConfigProvider {
         return proxyFactoryBean;
     }
 
+    @Bean
+    public DynamoDBMapperConfig dynamoDBMapperConfig() {
+        DynamoDBMapperConfig.Builder builder = new DynamoDBMapperConfig.Builder();
+        builder.withTableNameOverride(DynamoDBMapperConfig.TableNameOverride.
+                withTableNamePrefix(getDynamoDbPrefix()));
+        return builder.build();
+    }
+
+    @Bean
+    public ProxyFactoryBean amazonDynamoDbMapperProxy() {
+        ProxyFactoryBean proxyFactoryBean = new ProxyFactoryBean();
+
+        proxyFactoryBean.setTarget(dynamoDBMapper());
+        proxyFactoryBean.setInterceptorNames("retryInterceptor");
+
+        return proxyFactoryBean;
+    }
+
     @Bean(name = "dynamoDB")
     public AmazonDynamoDB amazonDynamoDB() {
         AmazonDynamoDB amazonDynamoDB = new AmazonDynamoDBClient(amazonCredentialsProvider());
         amazonDynamoDB.setRegion(Regions.getCurrentRegion());
         return amazonDynamoDB;
+    }
+
+    private DynamoDBMapper dynamoDBMapper() {
+        return new DynamoDBMapper(amazonDynamoDB(), dynamoDBMapperConfig());
     }
 
     private AmazonEC2 amazonEC2() {
@@ -84,5 +110,13 @@ public class AmazonConfigProvider {
             amazonS3.setRegion(current);
         }
         return amazonS3;
+    }
+
+    public static String getDynamoDbPrefix() {
+        return getDynamoDbPrefix(EC2MetadataUtils.getInstanceId());
+    }
+
+    public static String getDynamoDbPrefix(String instanceId) {
+        return "ENHANCEDSNAPSHOTS_" + instanceId + "_";
     }
 }
