@@ -25,14 +25,18 @@ public class Configuration {
     // sdfs settings
     private String sdfsVolumeName;
     private String sdfsMountPoint;
-    private String sdfsLocalCacheSize;
-    private String sdfsSize;
+    private int sdfsLocalCacheSize;
+    private int sdfsSize;
     private String sdfsConfigPath;
     private String sdfsBackupFileName;
 
     // amazon settings
     private String region;
     private String s3Bucket;
+
+    // Nginx
+    private String nginxCertPath;
+    private String nginxKeyPath;
 
     public String getSdfsVolumeName() {
         return sdfsVolumeName;
@@ -109,19 +113,19 @@ public class Configuration {
         this.restoreVolumeIopsPerGb = restoreVolumeIopsPerGb;
     }
 
-    public String getSdfsLocalCacheSize() {
+    public int getSdfsLocalCacheSize() {
         return sdfsLocalCacheSize;
     }
 
-    public void setSdfsLocalCacheSize(String sdfsLocalCacheSize) {
+    public void setSdfsLocalCacheSize(int sdfsLocalCacheSize) {
         this.sdfsLocalCacheSize = sdfsLocalCacheSize;
     }
 
-    public String getSdfsSize() {
+    public int getSdfsSize() {
         return sdfsSize;
     }
 
-    public void setSdfsSize(String sdfsSize) {
+    public void setSdfsSize(int sdfsSize) {
         this.sdfsSize = sdfsSize;
     }
 
@@ -197,4 +201,19 @@ public class Configuration {
         this.maxWaitTimeToDetachVolume = maxWaitTimeToDetachVolume;
     }
 
+    public String getNginxCertPath() {
+        return nginxCertPath;
+    }
+
+    public void setNginxCertPath(final String nginxCertPath) {
+        this.nginxCertPath = nginxCertPath;
+    }
+
+    public String getNginxKeyPath() {
+        return nginxKeyPath;
+    }
+
+    public void setNginxKeyPath(final String nginxKeyPath) {
+        this.nginxKeyPath = nginxKeyPath;
+    }
 }
