@@ -7,33 +7,19 @@ import com.amazonaws.services.dynamodbv2.datamodeling.DynamoDBTable;
 @DynamoDBTable(tableName = "Snapshots")
 public class SnapshotEntry {
 
-    private static final String DELIMITER = ":";
-
-    @DynamoDBAttribute
-    private String instanceId;
-
     @DynamoDBAttribute
     private String snapshotId;
 
-    @DynamoDBAttribute
+    @DynamoDBHashKey
     private String volumeId;
 
 
     public SnapshotEntry() {
     }
 
-    public SnapshotEntry(String instanceId, String snapshotId, String volumeId) {
-        this.instanceId = instanceId;
+    public SnapshotEntry(String snapshotId, String volumeId) {
         this.snapshotId = snapshotId;
         this.volumeId = volumeId;
-    }
-
-    public String getInstanceId() {
-        return instanceId;
-    }
-
-    public void setInstanceId(String instanceId) {
-        this.instanceId = instanceId;
     }
 
     public String getSnapshotId() {
@@ -50,20 +36,5 @@ public class SnapshotEntry {
 
     public void setVolumeId(String volumeId) {
         this.volumeId = volumeId;
-    }
-
-    @DynamoDBHashKey
-    public String getVolumeInstanceId() {
-        return volumeId + DELIMITER + instanceId;
-    }
-
-    public void setVolumeInstanceId(String id) {
-        String[] ids = id.split(DELIMITER);
-        volumeId = ids[0];
-        instanceId = ids[1];
-    }
-
-    public static final String getId(String volumeId, String instanceId) {
-        return volumeId + DELIMITER + instanceId;
     }
 }
