@@ -1,8 +1,8 @@
 package com.sungardas.enhancedsnapshots.rest;
 
-import com.sungardas.enhancedsnapshots.aws.dynamodb.repository.UserRepository;
-
 import javax.servlet.Filter;
+
+import com.sungardas.enhancedsnapshots.aws.dynamodb.repository.UserRepository;
 
 public interface RestAuthenticationFilter extends Filter {
     void setUserRepository(UserRepository userRepository);
