@@ -1,12 +1,20 @@
 package com.sungardas.enhancedsnapshots.rest;
 
+import java.util.Map;
+
+import javax.servlet.ServletContext;
+import javax.servlet.http.HttpServletRequest;
+
+import com.sungardas.enhancedsnapshots.components.ConfigurationMediator;
 import com.sungardas.enhancedsnapshots.components.WorkersDispatcher;
 import com.sungardas.enhancedsnapshots.dto.SystemConfiguration;
 import com.sungardas.enhancedsnapshots.rest.filters.FilterProxy;
 import com.sungardas.enhancedsnapshots.rest.utils.Constants;
-import com.sungardas.enhancedsnapshots.service.ConfigurationService;
+import com.sungardas.enhancedsnapshots.service.AWSCommunicationService;
 import com.sungardas.enhancedsnapshots.service.SDFSStateService;
+import com.sungardas.enhancedsnapshots.service.SystemService;
 import com.sungardas.enhancedsnapshots.service.UserService;
+
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
@@ -15,10 +23,6 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestMethod;
 import org.springframework.web.bind.annotation.RestController;
 import org.springframework.web.context.support.XmlWebApplicationContext;
-
-import javax.servlet.ServletContext;
-import javax.servlet.http.HttpServletRequest;
-import java.util.Map;
 
 
 @RestController
@@ -37,16 +41,22 @@ public class SystemController {
     private SDFSStateService sdfsStateService;
 
     @Autowired
-    private ConfigurationService configurationService;
+    private SystemService systemService;
 
     @Autowired
     private UserService userService;
+
+    @Autowired
+    private AWSCommunicationService awsCommunicationService;
 
     @Autowired
     private XmlWebApplicationContext applicationContext;
 
     @Autowired
     private WorkersDispatcher workersDispatcher;
+
+    @Autowired
+    private ConfigurationMediator configurationMediator;
 
     @RequestMapping(value = "/delete", method = RequestMethod.POST)
     public ResponseEntity<String> deleteService(@RequestBody InstanceId instanceId) {
@@ -55,7 +65,7 @@ public class SystemController {
         if (!userService.isAdmin(currentUser)) {
             return new ResponseEntity<>("{\"msg\":\"Only admin can delete service\"}", HttpStatus.FORBIDDEN);
         }
-        if (!configurationService.getConfigurationId().equals(instanceId.getInstanceId())) {
+        if (!configurationMediator.getConfigurationId().equals(instanceId.getInstanceId())) {
             return new ResponseEntity<>("{\"msg\":\"Provided instance ID is incorrect\"}", HttpStatus.FORBIDDEN);
         }
         refreshContext();
@@ -64,15 +74,35 @@ public class SystemController {
 
     @RequestMapping(method = RequestMethod.GET)
     public ResponseEntity<SystemConfiguration> getSystem() {
-        return new ResponseEntity<>(configurationService.getSystemConfiguration(), HttpStatus.OK);
+        return new ResponseEntity<>(systemService.getSystemConfiguration(), HttpStatus.OK);
     }
 
     @RequestMapping(method = RequestMethod.POST)
-    public ResponseEntity<String> setSystemProperties(@RequestBody SystemConfiguration.SystemProperties systemProperties) {
-        if (!checkIopsAreValid(systemProperties)) {
+    public ResponseEntity<String> updateSystemProperties(@RequestBody SystemConfiguration newConfiguration) {
+        SystemConfiguration currentConfiguration = systemService.getSystemConfiguration();
+        if (!checkIopsAreValid(newConfiguration.getSystemProperties())) {
             return new ResponseEntity<>("iops per GB can not be less than 1 and more than 30", HttpStatus.BAD_REQUEST);
         }
-        configurationService.setSystemProperties(systemProperties);
+        if (newConfiguration.getSdfs().getVolumeSize() > currentConfiguration.getSdfs().getMaxVolumeSize()) {
+            return new ResponseEntity<>("Volume size can not be more than " + currentConfiguration.getSdfs().getMaxVolumeSize(), HttpStatus.BAD_REQUEST);
+        }
+        if (newConfiguration.getSdfs().getSdfsLocalCacheSize() > currentConfiguration.getSdfs().getMaxSdfsLocalCacheSize()) {
+            return new ResponseEntity<>("Local cache size can not be more than " + currentConfiguration.getSdfs().getMaxSdfsLocalCacheSize(), HttpStatus.BAD_REQUEST);
+        }
+        boolean needToReconfigureSdfs = false;
+
+        if (configurationMediator.getSdfsVolumeSizeWithoutMeasureUnit() != newConfiguration.getSdfs().getVolumeSize()
+                && newConfiguration.getSdfs().getVolumeSize() > 0) {
+            sdfsStateService.expandSdfsVolume(newConfiguration.getSdfs().getVolumeSize() + configurationMediator.getVolumeSizeUnit());
+        }
+        if (configurationMediator.getSdfsLocalCacheSizeWithoutMeasureUnit() != newConfiguration.getSdfs().getSdfsLocalCacheSize()
+                && newConfiguration.getSdfs().getSdfsLocalCacheSize() > 0) {
+            needToReconfigureSdfs = true;
+        }
+        systemService.setSystemConfiguration(newConfiguration);
+        if (needToReconfigureSdfs) {
+            sdfsStateService.reconfigureAndRestartSDFS();
+        }
         return new ResponseEntity<>("", HttpStatus.OK);
     }
 
