@@ -1,8 +1,14 @@
 package com.sungardas.enhancedsnapshots.rest.filters;
 
 
-import javax.servlet.*;
 import java.io.IOException;
+
+import javax.servlet.Filter;
+import javax.servlet.FilterChain;
+import javax.servlet.FilterConfig;
+import javax.servlet.ServletException;
+import javax.servlet.ServletRequest;
+import javax.servlet.ServletResponse;
 
 
 public class FilterProxy implements Filter {
