@@ -1,16 +1,90 @@
 package com.sungardas.enhancedsnapshots.service;
 
+import com.sun.management.OperatingSystemMXBean;
+import com.sungardas.enhancedsnapshots.aws.dynamodb.model.BackupEntry;
+
+import java.io.File;
+import java.lang.management.ManagementFactory;
+import java.util.List;
+
 public interface SDFSStateService {
 
-    void backupState(String taskId);
+    // constant value
+    long BYTES_IN_GB = 1_073_741_824;
 
-    void restoreState();
 
+    /**
+     * Returns max sdfs volume size for current system in GB
+     * @param systemReservedRam in bytes
+     * @param volumeSizePerGbOfRam in GB
+     * @param sdfsReservedRam in bytes
+     * @return
+     */
+    static int getMaxVolumeSize(int systemReservedRam, int volumeSizePerGbOfRam,  int sdfsReservedRam) {
+        OperatingSystemMXBean osBean = (OperatingSystemMXBean) ManagementFactory.getOperatingSystemMXBean();
+        //Total RAM - RAM available for Tomcat - reserved
+        long totalRAM = osBean.getTotalPhysicalMemorySize() - Runtime.getRuntime().maxMemory() - systemReservedRam - sdfsReservedRam;
+        int maxVolumeSize = (int) (totalRAM / BYTES_IN_GB) * volumeSizePerGbOfRam;
+        return maxVolumeSize;
+    }
+
+
+    /**
+     * Returns count of GB which can be used to increase sdfs local cache
+     * @param systemReservedStorage reserved storage in bytes
+     * @return
+     */
+    static int getFreeStorageSpace(int systemReservedStorage) {
+        File file = new File("/");
+        int maxLocalCacheInGb = (int) ((file.getFreeSpace() - systemReservedStorage) / BYTES_IN_GB);
+        return maxLocalCacheInGb;
+    }
+
+    /**
+     * Return true when S3 bucket contains SDFS backup, false otherwise
+     */
     boolean containsSdfsMetadata(String sBucket);
 
     Long getBackupTime();
 
-    void shutdownSDFS(String size, String bucketName);
+    /**
+     * Reconfigure SDFS and restart
+     */
+    void reconfigureAndRestartSDFS();
 
-    void startupSDFS(String size, String bucketName, Boolean isRestore);
+    /**
+     * Restore SDFS from S3 bucket
+     */
+    void restoreSDFS();
+
+    /**
+     * Start SDFS if it is not running
+     */
+    void startSDFS();
+
+    /**
+     * Stop SDFS if it is not running
+     */
+    void stopSDFS();
+
+    /**
+     * Return true if SDFS is currently runnings, false otherwise
+     */
+    boolean sdfsIsAvailable();
+
+    /**
+     * Expand sdfs volume
+     */
+    void expandSdfsVolume(String newVolumeSize);
+
+    /**
+     * Sync local SDFS metadata with cloud
+     */
+    void cloudSync();
+
+    /**
+     * Returns list of real existing backups from SDFS mount point
+     * @return
+     */
+    List<BackupEntry> getBackupsFromSDFSMountPoint();
 }
