@@ -1,23 +1,25 @@
 package com.sungardas.enhancedsnapshots.service.impl;
 
+import java.util.ArrayList;
+import java.util.Collection;
+import java.util.List;
+import java.util.UUID;
+
+import javax.annotation.PostConstruct;
+
 import com.sungardas.enhancedsnapshots.aws.dynamodb.model.BackupEntry;
 import com.sungardas.enhancedsnapshots.aws.dynamodb.model.TaskEntry;
 import com.sungardas.enhancedsnapshots.aws.dynamodb.repository.BackupRepository;
 import com.sungardas.enhancedsnapshots.aws.dynamodb.repository.TaskRepository;
+import com.sungardas.enhancedsnapshots.components.ConfigurationMediator;
 import com.sungardas.enhancedsnapshots.exception.DataException;
 import com.sungardas.enhancedsnapshots.service.BackupService;
-import com.sungardas.enhancedsnapshots.service.ConfigurationService;
+
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
 import org.joda.time.DateTime;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
-
-import javax.annotation.PostConstruct;
-import java.util.ArrayList;
-import java.util.Collection;
-import java.util.List;
-import java.util.UUID;
 
 import static com.sungardas.enhancedsnapshots.aws.dynamodb.model.TaskEntry.TaskEntryType.DELETE;
 
@@ -29,7 +31,7 @@ public class BackupServiceImpl implements BackupService {
     private static final String BACKUP_FILE_EXT = ".backup";
 
     @Autowired
-    private ConfigurationService configurationService;
+    private ConfigurationMediator configurationMediator;
 
     @Autowired
     private BackupRepository backupRepository;
@@ -41,14 +43,14 @@ public class BackupServiceImpl implements BackupService {
 
     @PostConstruct
     private void init() {
-        instanceId = configurationService.getConfigurationId();
+        instanceId = configurationMediator.getConfigurationId();
     }
 
     @Override
     public void deleteBackup(String backupName, String user) {
         TaskEntry taskEntry = getDeleteTask(backupName + BACKUP_FILE_EXT, user, true);
-        if (taskRepository.findByVolumeAndTypeAndInstanceIdAndOptions(taskEntry.getVolume(),
-                taskEntry.getType(), instanceId, taskEntry.getOptions()).isEmpty()) {
+        if (taskRepository.findByVolumeAndTypeAndOptions(taskEntry.getVolume(),
+                taskEntry.getType(), taskEntry.getOptions()).isEmpty()) {
             taskRepository.save(taskEntry);
         } else {
             LOG.error("Task already exist: {}", taskEntry);
@@ -58,7 +60,7 @@ public class BackupServiceImpl implements BackupService {
 
     @Override
     public List<BackupEntry> getBackupList(String volumeId) {
-        return backupRepository.get(volumeId, instanceId);
+        return backupRepository.findByVolumeId(volumeId);
     }
 
     @Override
@@ -68,8 +70,8 @@ public class BackupServiceImpl implements BackupService {
 
         for (BackupEntry entry : backupEntries) {
             TaskEntry taskEntry = getDeleteTask(entry.getFileName(), user, false);
-            if (taskRepository.findByVolumeAndTypeAndInstanceIdAndOptions(taskEntry.getVolume(),
-                    taskEntry.getType(), instanceId, taskEntry.getOptions()).isEmpty()) {
+            if (taskRepository.findByVolumeAndTypeAndOptions(taskEntry.getVolume(),
+                    taskEntry.getType(), taskEntry.getOptions()).isEmpty()) {
 
                 tasks.add(getDeleteTask(entry.getFileName(), user, false));
             } else {
@@ -92,15 +94,14 @@ public class BackupServiceImpl implements BackupService {
         taskEntry.setId(UUID.randomUUID().toString());
         taskEntry.setVolume(volumeId);
         taskEntry.setType(DELETE.getType());
-        taskEntry.setInstanceId(instanceId);
         taskEntry.setStatus(TaskEntry.TaskEntryStatus.QUEUED.getStatus());
         taskEntry.setOptions(backupFile);
         taskEntry.setSchedulerName(user);
         taskEntry.setSchedulerTime(String.valueOf(DateTime.now().getMillis()));
         taskEntry.setPriority(1);
+        taskEntry.setWorker(configurationMediator.getConfigurationId());
 
         //TODO Remove hardcode
-        taskEntry.setWorker(taskEntry.getInstanceId());
         taskEntry.setSchedulerManual(schedulerManual);
         taskEntry.setRegular(false);
 
