@@ -1,15 +1,16 @@
 package com.sungardas.enhancedsnapshots.service.impl;
 
+import javax.annotation.PostConstruct;
+
 import com.amazonaws.regions.Regions;
 import com.amazonaws.services.dynamodbv2.AmazonDynamoDB;
 import com.amazonaws.services.dynamodbv2.datamodeling.DynamoDBMapper;
 import com.sungardas.enhancedsnapshots.aws.dynamodb.model.Configuration;
 import com.sungardas.enhancedsnapshots.aws.dynamodb.model.User;
+
 import org.apache.commons.codec.digest.DigestUtils;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Value;
-
-import javax.annotation.PostConstruct;
 
 public class CreateAppConfigurationDev {
 
@@ -36,7 +37,7 @@ public class CreateAppConfigurationDev {
     @Value("${enhancedsnapshots.default.polling.rate}")
     private int defaultPollingRate;
     @Value("${enhancedsnapshots.default.sdfs.local.cache.size}")
-    private String sdfsLocalCacheSize;
+    private int sdfsLocalCacheSize;
     @Value("${enhancedsnapshots.default.wait.time.before.new.sync}")
     private int defaultWaitTimeBeforeNewSyncWithAWS;
     @Value("${enhancedsnapshots.default.max.wait.time.to.detach.volume}")
@@ -76,6 +77,11 @@ public class CreateAppConfigurationDev {
         configuration.setWorkerDispatcherPollingRate(defaultPollingRate);
         configuration.setWaitTimeBeforeNewSyncWithAWS(defaultWaitTimeBeforeNewSyncWithAWS);
         configuration.setMaxWaitTimeToDetachVolume(defaultMaxWaitTimeToDetachVolume);
+        configuration.setS3Bucket("com.sungardas.enhancedsnapshots.dev");
+        configuration.setSdfsSize(500);
+        configuration.setSdfsVolumeName("awspool");
+        configuration.setSdfsMountPoint("/mnt/awspool");
+
         return configuration;
     }
 }
