@@ -2,8 +2,8 @@ package com.sungardas.enhancedsnapshots.service.impl;
 
 import com.sungardas.enhancedsnapshots.aws.dynamodb.model.SnapshotEntry;
 import com.sungardas.enhancedsnapshots.aws.dynamodb.repository.SnapshotRepository;
-import com.sungardas.enhancedsnapshots.service.ConfigurationService;
 import com.sungardas.enhancedsnapshots.service.SnapshotService;
+
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -17,21 +17,18 @@ public class SnapshotServiceImpl implements SnapshotService {
     @Autowired
     private SnapshotRepository snapshotRepository;
 
-    @Autowired
-    private ConfigurationService configurationService;
-
     @Override
-    public String getSnapshotId(String volumeId, String instancreId) {
-        LOG.info("Get snapshot id for volume {} and configuration {}", volumeId, instancreId);
+    public String getSnapshotId(String volumeId) {
+        LOG.info("Get snapshot id for volume {}", volumeId);
         try {
-            return snapshotRepository.findOne(SnapshotEntry.getId(volumeId, instancreId)).getSnapshotId();
+            return snapshotRepository.findOne(volumeId).getSnapshotId();
         } catch (Exception e) {
             return null;
         }
     }
 
     @Override
-    public void saveSnapshot(String volumeId, String instanceId, String snapshotId) {
-        snapshotRepository.save(new SnapshotEntry(instanceId, snapshotId, volumeId));
+    public void saveSnapshot(String volumeId, String snapshotId) {
+        snapshotRepository.save(new SnapshotEntry(snapshotId, volumeId));
     }
 }
