@@ -1,11 +1,27 @@
 package com.sungardas.enhancedsnapshots.service.impl;
 
+import java.util.HashMap;
+import java.util.List;
+import java.util.Map;
+import java.util.Set;
+import java.util.UUID;
+import java.util.concurrent.ScheduledFuture;
+import java.util.stream.Collectors;
+
+import javax.annotation.PostConstruct;
+
 import com.amazonaws.AmazonClientException;
 import com.sungardas.enhancedsnapshots.aws.dynamodb.model.TaskEntry;
 import com.sungardas.enhancedsnapshots.aws.dynamodb.repository.TaskRepository;
+import com.sungardas.enhancedsnapshots.components.ConfigurationMediator;
 import com.sungardas.enhancedsnapshots.dto.ExceptionDto;
 import com.sungardas.enhancedsnapshots.exception.EnhancedSnapshotsException;
-import com.sungardas.enhancedsnapshots.service.*;
+import com.sungardas.enhancedsnapshots.service.NotificationService;
+import com.sungardas.enhancedsnapshots.service.SchedulerService;
+import com.sungardas.enhancedsnapshots.service.Task;
+import com.sungardas.enhancedsnapshots.service.TaskService;
+import com.sungardas.enhancedsnapshots.service.VolumeService;
+
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
 import org.joda.time.DateTime;
@@ -16,12 +32,8 @@ import org.springframework.scheduling.TaskScheduler;
 import org.springframework.scheduling.support.CronTrigger;
 import org.springframework.stereotype.Service;
 
-import javax.annotation.PostConstruct;
-import java.util.*;
-import java.util.concurrent.ScheduledFuture;
-
 @Service
-@DependsOn("CreateAppConfiguration")
+@DependsOn("SystemService")
 public class SpringSchedulerService implements SchedulerService {
 
     private static final Logger LOG = LogManager.getLogger(SpringSchedulerService.class);
@@ -31,7 +43,7 @@ public class SpringSchedulerService implements SchedulerService {
     private TaskScheduler scheduler;
 
     @Autowired
-    private ConfigurationService configurationService;
+    private ConfigurationMediator configurationMediator;
 
     @Autowired
     private TaskRepository taskRepository;
@@ -52,8 +64,8 @@ public class SpringSchedulerService implements SchedulerService {
     @PostConstruct
     private void init() {
         try {
-            instanceId = configurationService.getConfigurationId();
-            List<TaskEntry> tasks = taskRepository.findByRegularAndInstanceId(Boolean.TRUE.toString(), instanceId);
+            instanceId = configurationMediator.getConfigurationId();
+            List<TaskEntry> tasks = taskRepository.findByRegular(Boolean.TRUE.toString());
             for (TaskEntry taskEntry : tasks) {
                 try {
                     addTask(taskEntry);
@@ -97,10 +109,7 @@ public class SpringSchedulerService implements SchedulerService {
 
     @Override
     public Set<String> getVolumeIdsWithSchedule() {
-        Set<String> result = new HashSet<>();
-        for (TaskEntry taskEntry : taskRepository.findByRegularAndInstanceIdAndEnabled(Boolean.TRUE.toString(), instanceId, Boolean.TRUE.toString())) {
-            result.add(taskEntry.getVolume());
-        }
+        Set<String> result = taskRepository.findByRegularAndEnabled(Boolean.TRUE.toString(), Boolean.TRUE.toString()).stream().map(TaskEntry::getVolume).collect(Collectors.toSet());
         return result;
     }
 
