@@ -3,42 +3,54 @@ package com.sungardas.init;
 import java.io.File;
 import java.io.FileWriter;
 import java.io.IOException;
-import java.lang.management.ManagementFactory;
 import java.nio.file.Paths;
 import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.List;
+import java.util.stream.Collectors;
 
 import javax.annotation.PostConstruct;
-import javax.validation.constraints.NotNull;
 
-import com.amazonaws.AmazonClientException;
 import com.amazonaws.AmazonServiceException;
 import com.amazonaws.auth.AWSCredentialsProvider;
-import com.amazonaws.auth.BasicAWSCredentials;
 import com.amazonaws.auth.InstanceProfileCredentialsProvider;
-import com.amazonaws.internal.StaticCredentialsProvider;
 import com.amazonaws.regions.Region;
 import com.amazonaws.regions.Regions;
 import com.amazonaws.services.dynamodbv2.AmazonDynamoDB;
 import com.amazonaws.services.dynamodbv2.AmazonDynamoDBClient;
 import com.amazonaws.services.dynamodbv2.datamodeling.DynamoDBMapper;
+import com.amazonaws.services.dynamodbv2.datamodeling.DynamoDBMapperConfig;
 import com.amazonaws.services.dynamodbv2.datamodeling.DynamoDBScanExpression;
 import com.amazonaws.services.dynamodbv2.document.DynamoDB;
 import com.amazonaws.services.dynamodbv2.document.Table;
-import com.amazonaws.services.dynamodbv2.model.*;
-import com.amazonaws.services.ec2.AmazonEC2Client;
-import com.amazonaws.services.s3.AmazonS3Client;
-import com.amazonaws.services.s3.model.*;
+
+import com.amazonaws.services.dynamodbv2.model.AttributeValue;
+import com.amazonaws.services.dynamodbv2.model.Condition;
+import com.amazonaws.services.dynamodbv2.model.CreateTableRequest;
+import com.amazonaws.services.dynamodbv2.model.ListTablesResult;
+import com.amazonaws.services.dynamodbv2.model.ProvisionedThroughput;
+import com.amazonaws.services.s3.AmazonS3;
+
+
+import com.amazonaws.services.s3.internal.BucketNameUtils;
+import com.amazonaws.services.s3.model.AmazonS3Exception;
+import com.amazonaws.services.s3.model.Bucket;
+import com.amazonaws.services.s3.model.ListObjectsRequest;
 import com.amazonaws.util.EC2MetadataUtils;
-import com.sun.management.UnixOperatingSystemMXBean;
+import com.sungardas.enhancedsnapshots.aws.AmazonConfigProvider;
+import com.sungardas.enhancedsnapshots.aws.dynamodb.model.BackupEntry;
 import com.sungardas.enhancedsnapshots.aws.dynamodb.model.Configuration;
+import com.sungardas.enhancedsnapshots.aws.dynamodb.model.RetentionEntry;
+import com.sungardas.enhancedsnapshots.aws.dynamodb.model.SnapshotEntry;
+import com.sungardas.enhancedsnapshots.aws.dynamodb.model.TaskEntry;
 import com.sungardas.enhancedsnapshots.aws.dynamodb.model.User;
 import com.sungardas.enhancedsnapshots.dto.InitConfigurationDto;
 import com.sungardas.enhancedsnapshots.dto.UserDto;
+import com.sungardas.enhancedsnapshots.dto.converter.BucketNameValidationDTO;
 import com.sungardas.enhancedsnapshots.dto.converter.UserDtoConverter;
 import com.sungardas.enhancedsnapshots.exception.ConfigurationException;
 import com.sungardas.enhancedsnapshots.exception.DataAccessException;
+import com.sungardas.enhancedsnapshots.service.SDFSStateService;
 
 import org.apache.commons.codec.digest.DigestUtils;
 import org.apache.commons.configuration2.PropertiesConfiguration;
@@ -47,6 +59,7 @@ import org.apache.commons.io.FileUtils;
 import org.apache.commons.io.FilenameUtils;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
+import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.scheduling.support.CronSequenceGenerator;
 import org.springframework.stereotype.Service;
@@ -57,17 +70,12 @@ import static com.amazonaws.services.dynamodbv2.model.ComparisonOperator.EQ;
 class InitConfigurationServiceImpl implements InitConfigurationService {
 
     private static final Logger LOG = LogManager.getLogger(InitConfigurationServiceImpl.class);
-    private static final long BYTES_IN_GB = 1_073_741_824;
-    private static final int SDFS_VOLUME_SIZE_IN_GB_PER_GB_OF_RAM = 2000;
-    private static final long SYSTEM_RESERVED_RAM_IN_BYTES = BYTES_IN_GB;
-    private static final long SDFS_RESERVED_RAM_IN_BYTES = BYTES_IN_GB;
 
     private static final String catalinaHomeEnvPropName = "catalina.home";
     private static final String confFolderName = "conf";
     private static final String propFileName = "EnhancedSnapshots.properties";
     private static final String DEFAULT_LOGIN = "admin@enhancedsnapshots";
 
-    private static final String ENHANCED_SNAPSHOT_BUCKET_PREFIX = "com.sungardas.enhancedsnapshots.";
     private static final String CANT_GET_ACCESS_DYNAMODB = "Can't get access to DynamoDB. Check policy list used for AWS user";
     private static final String CANT_GET_ACCESS_S3 = "Can't get access to S3. Check policy list used for AWS user";
 
@@ -81,12 +89,11 @@ class InitConfigurationServiceImpl implements InitConfigurationService {
     @Value("${enhancedsnapshots.retention.cron: -1}")
     private String retentionCronExpression;
     @Value("${enhancedsnapshots.polling.rate: -1}")
-    private int pollingRate;
+    private String pollingRate;
     @Value("${enhancedsnapshots.wait.time.before.new.sync: -1}")
-    private int waitTimeBeforeNewSync;
+    private String waitTimeBeforeNewSync;
     @Value("${enhancedsnapshots.max.wait.time.to.detach.volume: -1}")
-    private int maxWaitTimeToDetachVolume;
-    private int propertyNotProvided = -1;
+    private String maxWaitTimeToDetachVolume;
 
     @Value("${enhancedsnapshots.sdfs.min.size}")
     private String minVolumeSize;
@@ -115,35 +122,63 @@ class InitConfigurationServiceImpl implements InitConfigurationService {
     @Value("${enhancedsnapshots.default.polling.rate}")
     private int defaultPollingRate;
     @Value("${enhancedsnapshots.default.sdfs.local.cache.size}")
-    private String sdfsLocalCacheSize;
+    private int sdfsLocalCacheSize;
     @Value("${enhancedsnapshots.default.wait.time.before.new.sync}")
     private int defaultWaitTimeBeforeNewSyncWithAWS;
     @Value("${enhancedsnapshots.default.max.wait.time.to.detach.volume}")
     private int defaultMaxWaitTimeToDetachVolume;
     @Value("${enhancedsnapshots.default.sdfs.size}")
-    private String defaultVolumeSize;
+    private int defaultVolumeSize;
     @Value("${enhancedsnapshots.db.tables}")
     private String[] tables;
-    @Value("${amazon.s3.default.region}")
-    private String defaultS3Region;
     @Value("${enhancedsnapshots.default.sdfs.mount.point}")
     private String mountPoint;
     @Value("${enhancedsnapshots.default.sdfs.volume.name}")
     private String volumeName;
+
+    @Value("${enhancedsnapshots.bucket.name.prefix.001}")
+    private String enhancedSnapshotBucketPrefix001;
+    @Value("${enhancedsnapshots.bucket.name.prefix.002}")
+    private String enhancedSnapshotBucketPrefix002;
+
     @Value("${enhancedsnapshots.awscli.conf.path}")
     private String awscliConfPath;
     @Value("${enhancedsnapshots.awslogs.conf.path}")
     private String awslogsConfPath;
 
+    @Value("${enhancedsnapshots.nginx.cert.path}")
+    private String nginxCertPath;
+    @Value("${enhancedsnapshots.nginx.key.path}")
+    private String nginxKeyPath;
+    @Value("${enhancedsnapshots.db.read.capacity}")
+    private Long dbReadCapacity;
+    @Value("${enhancedsnapshots.db.write.capacity}")
+    private Long dbWriteCapacity;
+
+    @Value("${enhancedsnapshots.system.reserved.memory}")
+    private int systemReservedRam;
+    @Value("${enhancedsnapshots.sdfs.volume.size.per.1gb.ram}")
+    private int volumeSizePerGbOfRam;
+    @Value("${enhancedsnapshots.sdfs.reserved.ram}")
+    private int sdfsReservedRam;
+    @Value("${enhancedsnapshots.system.reserved.storage}")
+    private int systemReservedStorage;
+
+
+    @Autowired
+    private AmazonS3 amazonS3;
+
     private AWSCredentialsProvider credentialsProvider;
     private AmazonDynamoDB amazonDynamoDB;
     private DynamoDBMapper mapper;
-    private InitConfigurationDto initConfigurationDto;
     private String instanceId;
     private Region region;
 
     private UserDto userDto;
     private String adminPassword;
+
+    private List<String> tablesWithPrefix;
+    private String dbPrefix;
 
     @PostConstruct
     private void init() {
@@ -152,8 +187,11 @@ class InitConfigurationServiceImpl implements InitConfigurationService {
         region = Regions.getCurrentRegion();
         amazonDynamoDB = new AmazonDynamoDBClient(credentialsProvider);
         amazonDynamoDB.setRegion(region);
-
-        mapper = new DynamoDBMapper(amazonDynamoDB);
+        dbPrefix = AmazonConfigProvider.getDynamoDbPrefix();
+        DynamoDBMapperConfig config = new DynamoDBMapperConfig.Builder().withTableNameOverride(DynamoDBMapperConfig.TableNameOverride.
+                withTableNamePrefix(dbPrefix)).build();
+        mapper = new DynamoDBMapper(amazonDynamoDB, config);
+        tablesWithPrefix = Arrays.stream(tables).map(s -> dbPrefix.concat(s)).collect(Collectors.toList());
     }
 
     @Override
@@ -165,23 +203,16 @@ class InitConfigurationServiceImpl implements InitConfigurationService {
         }
     }
 
-    @Override
-    public void setInitConfigurationDto(InitConfigurationDto initConfigurationDto) {
-        this.initConfigurationDto = initConfigurationDto;
-    }
-
-    @Override
-    public void setCredentialsIfValid(@NotNull CredentialsDto credentials) {
-        validateCredentials(credentials.getAwsPublicKey(), credentials.getAwsSecretKey());
-        credentialsProvider = new StaticCredentialsProvider(new BasicAWSCredentials(credentials.getAwsPublicKey(), credentials.getAwsSecretKey()));
-    }
-
 
     /**
      *  Stores properties which can not be modified from UI to config file,
      *  changes in config file will be applied after system restart
      */
     public void storePropertiesEditableFromConfigFile() {
+        storePropertiesEditableFromConfigFile(defaultRetentionCronExpression, defaultPollingRate, defaultWaitTimeBeforeNewSyncWithAWS, defaultMaxWaitTimeToDetachVolume);
+    }
+
+    private void storePropertiesEditableFromConfigFile(String retentionCronExpression, int pollingRate, int waitTimeBeforeNewSync, int maxWaitTimeToDetachVolume) {
         File file = Paths.get(System.getProperty(catalinaHomeEnvPropName), confFolderName, propFileName).toFile();
         try {
             PropertiesConfiguration propertiesConfiguration = new PropertiesConfiguration();
@@ -193,19 +224,19 @@ class InitConfigurationServiceImpl implements InitConfigurationService {
 
             layout.setBlancLinesBefore(RETENTION_CRON_SCHEDULE, 1);
             layout.setComment(RETENTION_CRON_SCHEDULE, "Cron schedule for retention policy");
-            propertiesConfiguration.setProperty(RETENTION_CRON_SCHEDULE, defaultRetentionCronExpression);
+            propertiesConfiguration.setProperty(RETENTION_CRON_SCHEDULE, retentionCronExpression);
 
             layout.setBlancLinesBefore(WORKER_POLLING_RATE, 1);
             layout.setComment(WORKER_POLLING_RATE, "Polling rate to check whether there is some new task, in ms");
-            propertiesConfiguration.setProperty(WORKER_POLLING_RATE, Integer.toString(defaultPollingRate));
+            propertiesConfiguration.setProperty(WORKER_POLLING_RATE, Integer.toString(pollingRate));
 
             layout.setBlancLinesBefore(WAIT_TIME_BEFORE_NEXT_CHECK_IN_SECONDS, 1);
             layout.setComment(WAIT_TIME_BEFORE_NEXT_CHECK_IN_SECONDS, "Wait time before new sync of Snapshot/Volume with AWS data, in seconds");
-            propertiesConfiguration.setProperty(WAIT_TIME_BEFORE_NEXT_CHECK_IN_SECONDS, Integer.toString(defaultWaitTimeBeforeNewSyncWithAWS));
+            propertiesConfiguration.setProperty(WAIT_TIME_BEFORE_NEXT_CHECK_IN_SECONDS, Integer.toString(waitTimeBeforeNewSync));
 
             layout.setBlancLinesBefore(MAX_WAIT_TIME_VOLUME_TO_DETACH_IN_SECONDS, 1);
             layout.setComment(MAX_WAIT_TIME_VOLUME_TO_DETACH_IN_SECONDS, "Max wait time for volume to be detached, in seconds");
-            propertiesConfiguration.setProperty(MAX_WAIT_TIME_VOLUME_TO_DETACH_IN_SECONDS, Integer.toString(defaultMaxWaitTimeToDetachVolume));
+            propertiesConfiguration.setProperty(MAX_WAIT_TIME_VOLUME_TO_DETACH_IN_SECONDS, Integer.toString(maxWaitTimeToDetachVolume));
 
             propertiesConfiguration.write(new FileWriter(file));
             LOG.debug("Config file {} stored successfully.", file.getPath());
@@ -216,109 +247,67 @@ class InitConfigurationServiceImpl implements InitConfigurationService {
         }
     }
 
+    /**
+     * Create DB structure if required
+     * Store admin user if required
+     * Store configuration if required
+     * @param config
+     */
     public void createDBAndStoreSettings(final InitController.ConfigDto config) {
-        if (!requiredTablesExist()) { // check if tables corrupted
-            LOG.info("Initialization DB");
-            dropDbTables();
-            createDbAndStoreData(config);
-        } else {
-            storeAdminUserIfProvided();
-            if (!isConfigurationStored()) {
-                storeConfiguration(config);
-            }
-        }
-    }
-
-    private boolean isConfigurationStored() {
-        Configuration loadedConf = mapper.load(Configuration.class, EC2MetadataUtils.getInstanceId());
-        return loadedConf != null;
-    }
-
-    private void createDbAndStoreData(final InitController.ConfigDto config) {
+        // create tables if they do not exist
         createDbStructure();
         storeAdminUserIfProvided();
+        // store configuration if it does not exist
         storeConfiguration(config);
     }
 
     private void createDbStructure() throws ConfigurationException {
-        createTable("BackupList", 50L, 20L, "volumeId", "S", "fileName", "S");
-        createTable("Configurations", 10L, 10L, "configurationId", "S");
-        createTable("Retention", 50L, 20L, "volumeInstanceId", "S");
-        createTable("Tasks", 50L, 20L, "id", "S");
-        createTable("Snapshots", 50L, 20L, "volumeInstanceId", "S");
-        createTable("Users", 50L, 20L, "id", "S");
+        createTable(BackupEntry.class);
+        createTable(Configuration.class);
+        createTable(RetentionEntry.class);
+        createTable(TaskEntry.class);
+        createTable(SnapshotEntry.class);
+        createTable(User.class);
     }
 
-    private void createTable(String tableName, long readCapacityUnits, long writeCapacityUnits,
-            String hashKeyName, String hashKeyType) {
-        createTable(tableName, readCapacityUnits, writeCapacityUnits,
-                hashKeyName, hashKeyType, null, null);
-    }
+    private void createTable(Class tableClass) {
+        CreateTableRequest createTableRequest = mapper.generateCreateTableRequest(tableClass);
 
-    private void createTable(
-            String tableName, long readCapacityUnits, long writeCapacityUnits,
-            String hashKeyName, String hashKeyType,
-            String rangeKeyName, String rangeKeyType) {
-        DynamoDB dynamoDB = new DynamoDB(amazonDynamoDB);
+        createTableRequest.setProvisionedThroughput(new ProvisionedThroughput(dbReadCapacity, dbWriteCapacity));
+        if (tableExists(createTableRequest.getTableName())) {
+            LOG.info("Table {} already exists", createTableRequest.getTableName());
+            return;
+        }
         try {
-            ArrayList<KeySchemaElement> keySchema = new ArrayList<>();
-            keySchema.add(new KeySchemaElement()
-                    .withAttributeName(hashKeyName)
-                    .withKeyType(KeyType.HASH));
-
-            ArrayList<AttributeDefinition> attributeDefinitions = new ArrayList<>();
-            attributeDefinitions.add(new AttributeDefinition()
-                    .withAttributeName(hashKeyName)
-                    .withAttributeType(hashKeyType));
-
-            if (rangeKeyName != null) {
-                keySchema.add(new KeySchemaElement()
-                        .withAttributeName(rangeKeyName)
-                        .withKeyType(KeyType.RANGE));
-                attributeDefinitions.add(new AttributeDefinition()
-                        .withAttributeName(rangeKeyName)
-                        .withAttributeType(rangeKeyType));
-            }
-            CreateTableRequest request = new CreateTableRequest()
-                    .withTableName(tableName)
-                    .withKeySchema(keySchema)
-                    .withProvisionedThroughput(new ProvisionedThroughput()
-                            .withReadCapacityUnits(readCapacityUnits)
-                            .withWriteCapacityUnits(writeCapacityUnits));
-            request.setAttributeDefinitions(attributeDefinitions);
-            LOG.info("Issuing CreateTable request for " + tableName);
-            Table table = dynamoDB.createTable(request);
-            LOG.info("Waiting for " + tableName
-                    + " to be created...this may take a while...");
+            DynamoDB dynamoDB = new DynamoDB(amazonDynamoDB);
+            Table table = dynamoDB.createTable(createTableRequest);
+            LOG.info("Creating table {} ... ", createTableRequest.getTableName());
             table.waitForActive();
+            LOG.info("Table {} was created successfully.", createTableRequest.getTableName());
         } catch (Exception e) {
-            LOG.error("CreateTable request failed for " + tableName, e);
-            throw new ConfigurationException("CreateTable request failed for " + tableName, e);
+            LOG.error("Failed to create table {}. ", createTableRequest.getTableName());
+            LOG.error(e);
+            throw new ConfigurationException("Failed to create table" + createTableRequest.getTableName(), e);
         }
     }
+
 
     private void storeAdminUserIfProvided() {
         if (userDto != null && adminPassword != null) {
             userDto.setEmail(userDto.getEmail().toLowerCase());
             User userToCreate = UserDtoConverter.convert(userDto);
             userToCreate.setPassword(DigestUtils.sha512Hex(adminPassword));
-            userToCreate.setInstanceId(EC2MetadataUtils.getInstanceId());
             userToCreate.setRole("admin");
             mapper.save(userToCreate);
         }
     }
 
     private void storeConfiguration(final InitController.ConfigDto config) {
-        Configuration configuration = convertToWorkerConfiguration(initConfigurationDto, config);
-        mapper.save(configuration);
-    }
-
-    private Configuration convertToWorkerConfiguration(InitConfigurationDto dto, final InitController.ConfigDto config) {
         Configuration configuration = new Configuration();
         configuration.setConfigurationId(EC2MetadataUtils.getInstanceId());
         configuration.setEc2Region(Regions.getCurrentRegion().getName());
-        configuration.setSdfsMountPoint(dto.getSdfs().getMountPoint());
-        configuration.setSdfsVolumeName(dto.getSdfs().getVolumeName());
+        configuration.setSdfsMountPoint(mountPoint);
+        configuration.setSdfsVolumeName(volumeName);
         configuration.setS3Bucket(config.getBucketName());
         configuration.setSdfsSize(config.getVolumeSize());
 
@@ -327,19 +316,25 @@ class InitConfigurationServiceImpl implements InitConfigurationService {
         configuration.setRestoreVolumeType(restoreVolumeType);
         configuration.setTempVolumeIopsPerGb(tempVolumeIopsPerGb);
         configuration.setTempVolumeType(tempVolumeType);
-        configuration.setSdfsLocalCacheSize(sdfsLocalCacheSize);
         configuration.setAmazonRetryCount(amazonRetryCount);
         configuration.setAmazonRetrySleep(amazonRetrySleep);
         configuration.setMaxQueueSize(queueSize);
         configuration.setSdfsConfigPath(sdfsConfigPath);
+        configuration.setSdfsLocalCacheSize(sdfsLocalCacheSize);
         configuration.setSdfsBackupFileName(sdfsStateBackupFileName);
         configuration.setRetentionCronExpression(defaultRetentionCronExpression);
         configuration.setWorkerDispatcherPollingRate(defaultPollingRate);
         configuration.setWaitTimeBeforeNewSyncWithAWS(defaultWaitTimeBeforeNewSyncWithAWS);
         configuration.setMaxWaitTimeToDetachVolume(defaultMaxWaitTimeToDetachVolume);
-        return configuration;
+        configuration.setNginxCertPath(nginxCertPath);
+        configuration.setNginxKeyPath(nginxKeyPath);
+        // saving configuration to DB
+        mapper.save(configuration);
     }
 
+    /**
+     * Store properties changed in config file to DB
+     */
     public void syncSettingsInDbAndConfigFile() {
         // sync properties in DB with conf file
         boolean configurationChanged = false;
@@ -350,57 +345,42 @@ class InitConfigurationServiceImpl implements InitConfigurationService {
             loadedConf.setRetentionCronExpression(retentionCronExpression);
             configurationChanged = true;
         }
-        if (propertyProvided(pollingRate) && pollingRate != loadedConf.getWorkerDispatcherPollingRate()) {
+        if (correctPropertyProvided(pollingRate, WORKER_POLLING_RATE) && Integer.parseInt(pollingRate) != loadedConf.getWorkerDispatcherPollingRate()) {
             LOG.debug("Applying new polling rate to pick up new task {} ms.", pollingRate);
-            loadedConf.setWorkerDispatcherPollingRate(pollingRate);
+            loadedConf.setWorkerDispatcherPollingRate(Integer.parseInt(pollingRate));
             configurationChanged = true;
         }
-        if (propertyProvided(waitTimeBeforeNewSync) && waitTimeBeforeNewSync != loadedConf.getWaitTimeBeforeNewSyncWithAWS()) {
+        if (correctPropertyProvided(waitTimeBeforeNewSync, WAIT_TIME_BEFORE_NEXT_CHECK_IN_SECONDS)
+                && Integer.parseInt(waitTimeBeforeNewSync) != loadedConf.getWaitTimeBeforeNewSyncWithAWS()) {
             LOG.debug("Applying new wait time before new sync of Snapshot/Volume with AWS data {} seconds.", waitTimeBeforeNewSync);
-            loadedConf.setWaitTimeBeforeNewSyncWithAWS(waitTimeBeforeNewSync);
+            loadedConf.setWaitTimeBeforeNewSyncWithAWS(Integer.parseInt(waitTimeBeforeNewSync));
             configurationChanged = true;
         }
-        if (propertyProvided(maxWaitTimeToDetachVolume) && maxWaitTimeToDetachVolume != loadedConf.getMaxWaitTimeToDetachVolume()) {
+        if (correctPropertyProvided(maxWaitTimeToDetachVolume, MAX_WAIT_TIME_VOLUME_TO_DETACH_IN_SECONDS)
+                && Integer.parseInt(maxWaitTimeToDetachVolume) != loadedConf.getMaxWaitTimeToDetachVolume()) {
             LOG.debug("Applying new max wait time for volume to be detach {} seconds.", maxWaitTimeToDetachVolume);
-            loadedConf.setMaxWaitTimeToDetachVolume(maxWaitTimeToDetachVolume);
+            loadedConf.setMaxWaitTimeToDetachVolume(Integer.parseInt(maxWaitTimeToDetachVolume));
             configurationChanged = true;
         }
         if (configurationChanged) {
             LOG.debug("Storing updated settings from config file to DB.");
             mapper.save(loadedConf);
         }
+        // this is required to ensure that properties with invalid values in config file will be replaced with correct values
+        removeProperties();
+        storePropertiesEditableFromConfigFile(loadedConf.getRetentionCronExpression(), loadedConf.getWorkerDispatcherPollingRate(),
+                loadedConf.getWaitTimeBeforeNewSyncWithAWS(), loadedConf.getMaxWaitTimeToDetachVolume());
     }
 
+    /**
+     * Remove config file with properties
+     */
     @Override
     public void removeProperties() {
         File file = Paths.get(System.getProperty(catalinaHomeEnvPropName), confFolderName, propFileName).toFile();
         if (file.exists()) {
             file.delete();
         }
-    }
-
-    @Override
-    public boolean areCredentialsValid() {
-        AmazonEC2Client ec2Client = new AmazonEC2Client(credentialsProvider);
-        ec2Client.setRegion(region);
-        try {
-            ec2Client.describeRegions();
-            return true;
-        } catch (AmazonClientException e) {
-            LOG.warn("Provided AWS credentials are invalid.");
-            return false;
-        }
-    }
-
-    @Override
-    public boolean credentialsAreProvided() {
-        if (credentialsProvider.getCredentials() != null) {
-            validateCredentials(credentialsProvider.getCredentials().getAWSAccessKeyId(), credentialsProvider.getCredentials().getAWSSecretKey());
-            return true;
-        } else {
-            return false;
-        }
-
     }
 
     @Override
@@ -417,9 +397,8 @@ class InitConfigurationServiceImpl implements InitConfigurationService {
         ArrayList<InitConfigurationDto.S3> result = new ArrayList<>();
 
         try {
-            AmazonS3Client client = new AmazonS3Client(credentialsProvider);
-            List<Bucket> allBuckets = client.listBuckets();
-            String bucketName = ENHANCED_SNAPSHOT_BUCKET_PREFIX + instanceId;
+            List<Bucket> allBuckets = amazonS3.listBuckets();
+            String bucketName = enhancedSnapshotBucketPrefix002 + instanceId;
             result.add(new InitConfigurationDto.S3(bucketName, false));
 
             String currentLocation = region.toString();
@@ -428,8 +407,9 @@ class InitConfigurationServiceImpl implements InitConfigurationService {
             }
             for (Bucket bucket : allBuckets) {
                 try {
-                    if (bucket.getName().startsWith(ENHANCED_SNAPSHOT_BUCKET_PREFIX)) {
-                        String location = client.getBucketLocation(bucket.getName());
+
+                    if (bucket.getName().startsWith(enhancedSnapshotBucketPrefix001) || bucket.getName().startsWith(enhancedSnapshotBucketPrefix002)) {
+                        String location = amazonS3.getBucketLocation(bucket.getName());
 
                         // Because client.getBucketLocation(bucket.getName()) returns US if bucket is in us-east-1
                         if (!location.equalsIgnoreCase(currentLocation) && !location.equalsIgnoreCase("US")) {
@@ -438,7 +418,7 @@ class InitConfigurationServiceImpl implements InitConfigurationService {
 
                         ListObjectsRequest request = new ListObjectsRequest()
                                 .withBucketName(bucket.getName()).withPrefix(FilenameUtils.removeExtension(sdfsStateBackupFileName));
-                        if (client.listObjects(request).getObjectSummaries().size() > 0) {
+                        if (amazonS3.listObjects(request).getObjectSummaries().size() > 0) {
                             if (bucketName.equals(bucket.getName())) {
                                 result.get(0).setCreated(true);
                             } else {
@@ -455,14 +435,12 @@ class InitConfigurationServiceImpl implements InitConfigurationService {
             LOG.warn("Can't get access to S3");
             throw new DataAccessException(CANT_GET_ACCESS_S3, e);
         }
-       
 	return result;
-
     }
 
     @Override
     public InitConfigurationDto getInitConfigurationDto() {
-        initConfigurationDto = new InitConfigurationDto();
+        InitConfigurationDto initConfigurationDto = new InitConfigurationDto();
         initConfigurationDto.setDb(new InitConfigurationDto.DB());
         boolean isDbValid = requiredTablesExist();
         initConfigurationDto.getDb().setValid(isDbValid);
@@ -472,14 +450,18 @@ class InitConfigurationServiceImpl implements InitConfigurationService {
         InitConfigurationDto.SDFS sdfs = new InitConfigurationDto.SDFS();
         sdfs.setMountPoint(mountPoint);
         sdfs.setVolumeName(volumeName);
-        int maxVolumeSize = getMaxVolumeSize();
+        int maxVolumeSize = SDFSStateService.getMaxVolumeSize(systemReservedRam, volumeSizePerGbOfRam,  sdfsReservedRam);
         sdfs.setMaxVolumeSize(String.valueOf(maxVolumeSize));
-        sdfs.setVolumeSize(String.valueOf(Math.min(maxVolumeSize, Integer.parseInt(defaultVolumeSize))));
+        sdfs.setVolumeSize(String.valueOf(Math.min(maxVolumeSize, defaultVolumeSize)));
         sdfs.setMinVolumeSize(minVolumeSize);
         sdfs.setCreated(sdfsAlreadyExists());
+        sdfs.setSdfsLocalCacheSize(sdfsLocalCacheSize);
+        sdfs.setMinSdfsLocalCacheSize(1);
+        sdfs.setMaxSdfsLocalCacheSize(SDFSStateService.getFreeStorageSpace(systemReservedStorage));
 
         initConfigurationDto.setS3(getBucketsWithSdfsMetadata());
         initConfigurationDto.setSdfs(sdfs);
+        initConfigurationDto.setImmutableBucketNamePrefix(enhancedSnapshotBucketPrefix002);
         return initConfigurationDto;
     }
 
@@ -489,9 +471,10 @@ class InitConfigurationServiceImpl implements InitConfigurationService {
         try {
             ListTablesResult listResult = amazonDynamoDB.listTables();
             List<String> tableNames = listResult.getTableNames();
+            boolean present = tableNames.containsAll(tablesWithPrefix);
             LOG.info("List db structure: {}", tableNames.toArray());
-            LOG.info("Check db structure is present: {}", tableNames.containsAll(Arrays.asList(tables)));
-            return tableNames.containsAll(Arrays.asList(tables));
+            LOG.info("Check db structure is present: {}", present);
+            return present;
         } catch (AmazonServiceException e) {
             LOG.warn("Can't get a list of existed tables", e);
             throw new DataAccessException(CANT_GET_ACCESS_DYNAMODB, e);
@@ -520,15 +503,6 @@ class InitConfigurationServiceImpl implements InitConfigurationService {
         return configf.exists() && mountPointf.exists();
     }
 
-    private void validateCredentials(String accessKey, String secretKey) {
-        if (accessKey == null || accessKey.isEmpty()) {
-            throw new ConfigurationException("Empty AWS AccessKey");
-        }
-        if (secretKey == null || secretKey.isEmpty()) {
-            throw new ConfigurationException("Empty AWS SecretKey");
-        }
-    }
-
     private File getPropertyFile() {
         return Paths.get(System.getProperty(catalinaHomeEnvPropName), confFolderName, propFileName).toFile();
     }
@@ -550,11 +524,10 @@ class InitConfigurationServiceImpl implements InitConfigurationService {
     }
 
     @Override
-    public void validateVolumeSize(final String volumeSize) {
-        int size = Integer.parseInt(volumeSize);
+    public void validateVolumeSize(final int volumeSize) {
         int min = Integer.parseInt(minVolumeSize);
-        int max = getMaxVolumeSize();
-        if (size < min || size > max) {
+        int max = SDFSStateService.getMaxVolumeSize(systemReservedRam, volumeSizePerGbOfRam,  sdfsReservedRam);
+        if (volumeSize < min || volumeSize > max) {
             throw new ConfigurationException("Invalid volume size");
         }
     }
@@ -569,41 +542,23 @@ class InitConfigurationServiceImpl implements InitConfigurationService {
         FileUtils.writeLines(file, Arrays.asList(lines));
     }
 
-    public int getMaxVolumeSize() {
-        UnixOperatingSystemMXBean osBean = (UnixOperatingSystemMXBean) ManagementFactory.getOperatingSystemMXBean();
-        //Total RAM - RAM available for Tomcat - reserved
-        long totalRAM = osBean.getFreePhysicalMemorySize() - Runtime.getRuntime().maxMemory() - SYSTEM_RESERVED_RAM_IN_BYTES - SDFS_RESERVED_RAM_IN_BYTES;
-        int maxVolumeSize = (int) (totalRAM / BYTES_IN_GB) * SDFS_VOLUME_SIZE_IN_GB_PER_GB_OF_RAM;
-        return maxVolumeSize;
-    }
-
-    private void dropDbTables() {
-        DynamoDB dynamoDB = new DynamoDB(amazonDynamoDB);
+    private boolean tableExists(String tableName) {
         ListTablesResult listResult = amazonDynamoDB.listTables();
         List<String> tableNames = listResult.getTableNames();
-        for (String tableToDelete : tables) {
-            if (tableNames.contains(tableToDelete)) {
-                try {
-                    Table table = dynamoDB.getTable(tableToDelete);
-                    table.delete();
-                    table.waitForDelete();
-                    LOG.info("Table {} was removed successfully.", table);
-                } catch (ResourceNotFoundException e) {
-                    // Skip exception if resource not found
-                } catch (AmazonServiceException tableNotFoundOrCredError) {
-                    LOG.warn("Failed to remove table {}", tableToDelete);
-                    throw new ConfigurationException("Can't delete tables. check AWS credentials");
-                } catch (InterruptedException e) {
-                    throw new ConfigurationException(e);
-                }
-            }
-        }
+        return tableNames.contains(tableName);
     }
 
-    // when applications starts for the first time property file in conf directory does not exist yet
-    // in case user removes some properties from file this check will help to avoid unwanted exceptions
-    private boolean propertyProvided(int property) {
-        return property != propertyNotProvided;
+    // in case user removes some properties from file or provided incorrect int values this check will help to avoid unwanted exceptions
+    private boolean correctPropertyProvided(String propertyValue, String propertyName) {
+        try {
+            int newProperty = Integer.parseInt(propertyValue);
+            if (newProperty > 0) {
+                return true;
+            }
+        } catch (Exception igrone) {
+        }
+        LOG.warn("Incorrect value {} for property {}. Changes will not bve applied.", propertyValue, propertyName);
+        return false;
     }
 
     private boolean cronExpressionIsValid(String cronExpression) {
@@ -615,7 +570,54 @@ class InitConfigurationServiceImpl implements InitConfigurationService {
             LOG.warn("Provided cron expression {} is invalid. Changes will not be applied", cronExpression);
             return false;
         }
+    }
 
+    public BucketNameValidationDTO validateBucketName(String bucketName) {
+        if (!bucketName.startsWith(enhancedSnapshotBucketPrefix002)) {
+            return new BucketNameValidationDTO(false, "Bucket name should start with " + enhancedSnapshotBucketPrefix002);
+        }
+        if (amazonS3.doesBucketExist(bucketName)) {
+            // check whether we own this bucket
+            List<Bucket> buckets = amazonS3.listBuckets();
+            for (Bucket bucket : buckets) {
+                if (bucket.getName().equals(bucketName)) {
+                    return new BucketNameValidationDTO(true, "");
+                }
+            }
+            return new BucketNameValidationDTO(false, "The requested bucket name is not available.Please select a different name.");
+        }
+        try {
+            BucketNameUtils.validateBucketName(bucketName);
+            return new BucketNameValidationDTO(true, "");
+        } catch (IllegalArgumentException e) {
+            return new BucketNameValidationDTO(false, e.getMessage());
+        }
+    }
+
+    @Override
+    public void createBucket(String bucketName) {
+        if (!bucketExists(bucketName)) {
+            LOG.info("Creating bucket {} in {} region", bucketName, region);
+            // AWS throws exception when trying create bucket in US_EAST_1
+            if (region.getName().equalsIgnoreCase(Regions.US_EAST_1.getName())) {
+                amazonS3.createBucket(bucketName);
+                return;
+            }
+            amazonS3.createBucket(bucketName, region.getName());
+        }
+    }
+
+    // there is bug at US_EAST_1 AWS S3 endpoint, we should not check buckets existence with it
+    private boolean bucketExists(String bucketName) {
+        boolean result;
+        if (region.getName().equalsIgnoreCase(Regions.US_EAST_1.getName())) {
+            amazonS3.setRegion(Region.getRegion(Regions.EU_CENTRAL_1));
+            result = amazonS3.doesBucketExist(bucketName);
+            amazonS3.setRegion(region);
+        } else {
+            result = amazonS3.doesBucketExist(bucketName);
+        }
+        return result;
     }
 
 }
