@@ -1,13 +1,12 @@
 package com.sungardas.init;
 
-import java.util.Arrays;
-
 import javax.annotation.PostConstruct;
 
 import com.amazonaws.AmazonClientException;
 import com.sungardas.enhancedsnapshots.aws.dynamodb.model.User;
 import com.sungardas.enhancedsnapshots.aws.dynamodb.repository.UserRepository;
 import com.sungardas.enhancedsnapshots.dto.InitConfigurationDto;
+import com.sungardas.enhancedsnapshots.dto.converter.BucketNameValidationDTO;
 import com.sungardas.enhancedsnapshots.exception.ConfigurationException;
 import com.sungardas.enhancedsnapshots.exception.EnhancedSnapshotsException;
 import com.sungardas.enhancedsnapshots.rest.RestAuthenticationFilter;
@@ -21,30 +20,23 @@ import org.springframework.context.ApplicationContext;
 import org.springframework.context.ApplicationContextAware;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.ExceptionHandler;
-import org.springframework.web.bind.annotation.RequestBody;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RequestMethod;
-import org.springframework.web.bind.annotation.ResponseBody;
-import org.springframework.web.bind.annotation.ResponseStatus;
-import org.springframework.web.bind.annotation.RestController;
+import org.springframework.web.bind.annotation.*;
 import org.springframework.web.context.support.XmlWebApplicationContext;
 
 import static org.springframework.http.HttpStatus.INTERNAL_SERVER_ERROR;
-import static org.springframework.http.HttpStatus.OK;
 
 
 @RestController
 class InitController implements ApplicationContextAware {
 
     private static final Logger LOG = LogManager.getLogger(InitController.class);
-    private static final String GB_UNIT = "GB";
 
     @Autowired
     private FilterProxy filterProxy;
 
     @Autowired
     private InitConfigurationService initConfigurationService;
+    private InitConfigurationDto configurationDto;
 
     @Autowired
     private XmlWebApplicationContext applicationContext;
@@ -61,7 +53,6 @@ class InitController implements ApplicationContextAware {
             refreshContext();
         } else {
             initConfigurationService.configureAWSLogAgent();
-
         }
     }
 
@@ -73,6 +64,11 @@ class InitController implements ApplicationContextAware {
         return exception;
     }
 
+    //TODO: This should be removed, for dev mode only
+    @RequestMapping(value = "/configuration/awscreds", method = RequestMethod.GET)
+    public ResponseEntity<String> getAwsCredentialsInfo() {
+        return new ResponseEntity<>("{\"contains\": true}", HttpStatus.OK);
+    }
 
     @ExceptionHandler(value = {Exception.class, AmazonClientException.class})
     @ResponseBody
@@ -96,58 +92,46 @@ class InitController implements ApplicationContextAware {
         }
     }
 
-    @RequestMapping(value = "/configuration/awscreds", method = RequestMethod.POST)
-    public ResponseEntity<String> setAwsCredential(@RequestBody CredentialsDto credentials) {
-        initConfigurationService.setCredentialsIfValid(credentials);
-        LOG.info("provided aws keys");
-        return new ResponseEntity<>(OK);
-    }
-
-    @RequestMapping(value = "/configuration/awscreds", method = RequestMethod.GET)
-    public ResponseEntity<String> getAwsCredentialsInfo() {
-        if (initConfigurationService.credentialsAreProvided()) {
-            return new ResponseEntity<>("{\"contains\": true}", HttpStatus.OK);
-        } else {
-            return new ResponseEntity<>("{\"contains\": false}", HttpStatus.OK);
-        }
-    }
-
-
     @RequestMapping(value = "/configuration/current", method = RequestMethod.GET)
     public ResponseEntity<InitConfigurationDto> getConfiguration() {
-        return new ResponseEntity<>(initConfigurationService.getInitConfigurationDto(), HttpStatus.OK);
+        return new ResponseEntity<>(getInitConfigurationDTO(), HttpStatus.OK);
     }
 
 
     @RequestMapping(value = "/configuration/current", method = RequestMethod.POST)
     public ResponseEntity<String> setConfiguration(@RequestBody ConfigDto config) {
-        if (initConfigurationService.areCredentialsValid()) {
-            InitConfigurationDto initConfigurationDto = initConfigurationService.getInitConfigurationDto();
-            if (!initConfigurationDto.getDb().isValid()) {
-                if (config.getUser() == null) {
-                    throw new ConfigurationException("Please create default user");
-                }
-                initConfigurationService.setUser(config.getUser());
+        InitConfigurationDto initConfigurationDto = getInitConfigurationDTO();
+        if (!initConfigurationDto.getDb().isValid()) {
+            if (config.getUser() == null) {
+                throw new ConfigurationException("Please create default user");
             }
-            if (config.getUser() != null) {
-                initConfigurationService.setUser(config.getUser());
-            }
-            initConfigurationService.validateVolumeSize(config.getVolumeSize());
-            initConfigurationDto.getSdfs().setVolumeSize(config.getVolumeSize() + GB_UNIT);
-            initConfigurationDto.setS3(Arrays.asList(new InitConfigurationDto.S3(config.getBucketName(), false)));
-            initConfigurationService.setInitConfigurationDto(initConfigurationDto);
-            initConfigurationService.storePropertiesEditableFromConfigFile();
-            initConfigurationService.createDBAndStoreSettings(config);
-            try {
-                refreshContext();
-            } catch (Exception e) {
-                initConfigurationService.removeProperties();
-                throw e;
-            }
-            return new ResponseEntity<>("", HttpStatus.OK);
-        } else {
-            throw new ConfigurationException("AWS configuration invalid");
+            initConfigurationService.setUser(config.getUser());
         }
+        if (config.getUser() != null) {
+            initConfigurationService.setUser(config.getUser());
+        }
+        initConfigurationService.validateVolumeSize(config.getVolumeSize());
+        try {
+            // we need to ensure before context refresh that provided bucket name is valid
+            initConfigurationService.createBucket(config.getBucketName());
+        } catch (IllegalArgumentException e) {
+            LOG.warn("Failed to create bucket {}", config.getBucketName());
+            return new ResponseEntity<>(e.getMessage(), HttpStatus.UNPROCESSABLE_ENTITY);
+        }
+        initConfigurationService.storePropertiesEditableFromConfigFile();
+        initConfigurationService.createDBAndStoreSettings(config);
+        try {
+            refreshContext();
+        } catch (Exception e) {
+            initConfigurationService.removeProperties();
+            throw e;
+        }
+        return new ResponseEntity<>("", HttpStatus.OK);
+    }
+
+    @RequestMapping(value = "/configuration/bucket/{name:.+}", method = RequestMethod.GET)
+    public ResponseEntity<BucketNameValidationDTO> validateBucketName(@PathVariable("name") String bucketName) {
+        return new ResponseEntity<>(initConfigurationService.validateBucketName(bucketName), HttpStatus.OK);
     }
 
     @Override
@@ -174,13 +158,13 @@ class InitController implements ApplicationContextAware {
     static class ConfigDto {
         private User user;
         private String bucketName;
-        private String volumeSize;
+        private int volumeSize;
 
-        public String getVolumeSize() {
+        public int getVolumeSize() {
             return volumeSize;
         }
 
-        public void setVolumeSize(final String volumeSize) {
+        public void setVolumeSize(final int volumeSize) {
             this.volumeSize = volumeSize;
         }
 
@@ -199,5 +183,12 @@ class InitController implements ApplicationContextAware {
         public void setBucketName(String bucketName) {
             this.bucketName = bucketName;
         }
+    }
+
+    private InitConfigurationDto getInitConfigurationDTO() {
+        if (configurationDto == null) {
+            configurationDto = initConfigurationService.getInitConfigurationDto();
+        }
+        return configurationDto;
     }
 }
