@@ -57,13 +57,13 @@ public class RetentionServiceImplTest {
 
     @Test
     public void applyTest() {
-        BackupEntry entry1 = new BackupEntry("volumeId 1", "fileName1", new DateTime().plusHours(1).getMillis() + "", "1", COMPLETED, "", "", "", "", "");
-        BackupEntry entry2 = new BackupEntry("volumeId 1", "fileName2", new DateTime().plusDays(1).getMillis() + "", "1", COMPLETED, "", "", "", "", "");
-        BackupEntry entry3 = new BackupEntry("volumeId 1", "fileName3", new DateTime().plusDays(2).getMillis() + "", "1", COMPLETED, "", "", "", "", "");
-        when(backupRepository.findAll(null)).thenReturn(Arrays.asList(entry1, entry2, entry3));
+        BackupEntry entry1 = new BackupEntry("volumeId 1", "fileName1", new DateTime().plusHours(1).getMillis() + "", "1", COMPLETED, "", "", "", "");
+        BackupEntry entry2 = new BackupEntry("volumeId 1", "fileName2", new DateTime().plusDays(1).getMillis() + "", "1", COMPLETED, "", "", "", "");
+        BackupEntry entry3 = new BackupEntry("volumeId 1", "fileName3", new DateTime().plusDays(2).getMillis() + "", "1", COMPLETED, "", "", "", "");
+        when(backupRepository.findAll()).thenReturn(Arrays.asList(entry1, entry2, entry3));
 
         RetentionEntry retentionEntry = new RetentionEntry("volumeId 1", 2, 1, 1);
-        when(retentionRepository.findByInstanceId(null)).thenReturn(Arrays.asList(retentionEntry));
+        when(retentionRepository.findAll()).thenReturn(Arrays.asList(retentionEntry));
 
         retentionService.apply();
 
@@ -86,13 +86,13 @@ public class RetentionServiceImplTest {
 
     @Test
     public void applyCountRetentionPolicyTest() {
-        BackupEntry entry1 = new BackupEntry("volumeId 1", "fileName1", new DateTime().plusHours(1).getMillis() + "", "1", COMPLETED, "", "", "", "", "");
-        BackupEntry entry2 = new BackupEntry("volumeId 1", "fileName2", new DateTime().plusDays(1).getMillis() + "", "1", COMPLETED, "", "", "", "", "");
-        BackupEntry entry3 = new BackupEntry("volumeId 1", "fileName3", new DateTime().plusDays(2).getMillis() + "", "1", COMPLETED, "", "", "", "", "");
-        when(backupRepository.findAll(null)).thenReturn(Arrays.asList(entry1, entry2, entry3));
+        BackupEntry entry1 = new BackupEntry("volumeId 1", "fileName1", new DateTime().plusHours(1).getMillis() + "", "1", COMPLETED, "", "", "", "");
+        BackupEntry entry2 = new BackupEntry("volumeId 1", "fileName2", new DateTime().plusDays(1).getMillis() + "", "1", COMPLETED, "", "", "", "");
+        BackupEntry entry3 = new BackupEntry("volumeId 1", "fileName3", new DateTime().plusDays(2).getMillis() + "", "1", COMPLETED, "", "", "", "");
+        when(backupRepository.findAll()).thenReturn(Arrays.asList(entry1, entry2, entry3));
 
         RetentionEntry retentionEntry = new RetentionEntry("volumeId 1", 0, 1, 0);
-        when(retentionRepository.findByInstanceId(null)).thenReturn(Arrays.asList(retentionEntry));
+        when(retentionRepository.findAll()).thenReturn(Arrays.asList(retentionEntry));
 
         retentionService.apply();
 
@@ -113,13 +113,13 @@ public class RetentionServiceImplTest {
 
     @Test
     public void applySizeRetentionPolicyTest1() {
-        BackupEntry entry1 = new BackupEntry("volumeId 1", "fileName1", new DateTime().plusHours(1).getMillis() + "", "1073741824", COMPLETED, "", "", "", "", "");
-        BackupEntry entry2 = new BackupEntry("volumeId 1", "fileName2", new DateTime().plusDays(1).getMillis() + "", "1073741824", COMPLETED, "", "", "", "", "");
-        BackupEntry entry3 = new BackupEntry("volumeId 1", "fileName3", new DateTime().plusDays(2).getMillis() + "", "1073741824", COMPLETED, "", "", "", "", "");
-        when(backupRepository.findAll(null)).thenReturn(Arrays.asList(entry1, entry2, entry3));
+        BackupEntry entry1 = new BackupEntry("volumeId 1", "fileName1", new DateTime().plusHours(1).getMillis() + "", "1073741824", COMPLETED, "", "", "", "");
+        BackupEntry entry2 = new BackupEntry("volumeId 1", "fileName2", new DateTime().plusDays(1).getMillis() + "", "1073741824", COMPLETED, "", "", "", "");
+        BackupEntry entry3 = new BackupEntry("volumeId 1", "fileName3", new DateTime().plusDays(2).getMillis() + "", "1073741824", COMPLETED, "", "", "", "");
+        when(backupRepository.findAll()).thenReturn(Arrays.asList(entry1, entry2, entry3));
 
         RetentionEntry retentionEntry = new RetentionEntry("volumeId 1", 1, 0, 0);
-        when(retentionRepository.findByInstanceId(null)).thenReturn(Arrays.asList(retentionEntry));
+        when(retentionRepository.findAll()).thenReturn(Arrays.asList(retentionEntry));
 
         retentionService.apply();
 
@@ -140,13 +140,13 @@ public class RetentionServiceImplTest {
 
     @Test
     public void applySizeRetentionPolicyTest2() {
-        BackupEntry entry1 = new BackupEntry("volumeId 1", "fileName1", new DateTime().plusHours(1).getMillis() + "", "1073741824", COMPLETED, "", "", "", "", "");
-        BackupEntry entry2 = new BackupEntry("volumeId 1", "fileName2", new DateTime().plusDays(1).getMillis() + "", "1073741824", COMPLETED, "", "", "", "", "");
-        BackupEntry entry3 = new BackupEntry("volumeId 1", "fileName3", new DateTime().plusDays(2).getMillis() + "", "1073741824", COMPLETED, "", "", "", "", "");
-        when(backupRepository.findAll(null)).thenReturn(Arrays.asList(entry1, entry2, entry3));
+        BackupEntry entry1 = new BackupEntry("volumeId 1", "fileName1", new DateTime().plusHours(1).getMillis() + "", "1073741824", COMPLETED, "", "", "", "");
+        BackupEntry entry2 = new BackupEntry("volumeId 1", "fileName2", new DateTime().plusDays(1).getMillis() + "", "1073741824", COMPLETED, "", "", "", "");
+        BackupEntry entry3 = new BackupEntry("volumeId 1", "fileName3", new DateTime().plusDays(2).getMillis() + "", "1073741824", COMPLETED, "", "", "", "");
+        when(backupRepository.findAll()).thenReturn(Arrays.asList(entry1, entry2, entry3));
 
         RetentionEntry retentionEntry = new RetentionEntry("volumeId 1", 2, 0, 0);
-        when(retentionRepository.findByInstanceId(null)).thenReturn(Arrays.asList(retentionEntry));
+        when(retentionRepository.findAll()).thenReturn(Arrays.asList(retentionEntry));
 
         retentionService.apply();
 
@@ -166,13 +166,13 @@ public class RetentionServiceImplTest {
 
     @Test
     public void applyDayRetentionPolicyTest1() {
-        BackupEntry entry1 = new BackupEntry("volumeId 1", "fileName1", new DateTime().minusDays(2).getMillis() + "", "1", COMPLETED, "", "", "", "", "");
-        BackupEntry entry2 = new BackupEntry("volumeId 1", "fileName2", new DateTime().plusHours(23).getMillis() + "", "1", COMPLETED, "", "", "", "", "");
-        BackupEntry entry3 = new BackupEntry("volumeId 1", "fileName3", new DateTime().plusDays(2).getMillis() + "", "1", COMPLETED, "", "", "", "", "");
-        when(backupRepository.findAll(null)).thenReturn(Arrays.asList(entry1, entry2, entry3));
+        BackupEntry entry1 = new BackupEntry("volumeId 1", "fileName1", new DateTime().minusDays(2).getMillis() + "", "1", COMPLETED, "", "", "", "");
+        BackupEntry entry2 = new BackupEntry("volumeId 1", "fileName2", new DateTime().plusHours(23).getMillis() + "", "1", COMPLETED, "", "", "", "");
+        BackupEntry entry3 = new BackupEntry("volumeId 1", "fileName3", new DateTime().plusDays(2).getMillis() + "", "1", COMPLETED, "", "", "", "");
+        when(backupRepository.findAll()).thenReturn(Arrays.asList(entry1, entry2, entry3));
 
         RetentionEntry retentionEntry = new RetentionEntry("volumeId 1", 0, 0, 1);
-        when(retentionRepository.findByInstanceId(null)).thenReturn(Arrays.asList(retentionEntry));
+        when(retentionRepository.findAll()).thenReturn(Arrays.asList(retentionEntry));
 
         retentionService.apply();
 
@@ -192,13 +192,13 @@ public class RetentionServiceImplTest {
 
     @Test
     public void applyDayRetentionPolicyTest2() {
-        BackupEntry entry1 = new BackupEntry("volumeId 1", "fileName1", new DateTime().minusDays(2).getMillis() + "", "1", COMPLETED, "", "", "", "", "");
-        BackupEntry entry2 = new BackupEntry("volumeId 1", "fileName2", new DateTime().minusDays(23).getMillis() + "", "1", COMPLETED, "", "", "", "", "");
-        BackupEntry entry3 = new BackupEntry("volumeId 1", "fileName3", new DateTime().plusDays(2).getMillis() + "", "1", COMPLETED, "", "", "", "", "");
-        when(backupRepository.findAll(null)).thenReturn(Arrays.asList(entry1, entry2, entry3));
+        BackupEntry entry1 = new BackupEntry("volumeId 1", "fileName1", new DateTime().minusDays(2).getMillis() + "", "1", COMPLETED, "", "", "", "");
+        BackupEntry entry2 = new BackupEntry("volumeId 1", "fileName2", new DateTime().minusDays(23).getMillis() + "", "1", COMPLETED, "", "", "", "");
+        BackupEntry entry3 = new BackupEntry("volumeId 1", "fileName3", new DateTime().plusDays(2).getMillis() + "", "1", COMPLETED, "", "", "", "");
+        when(backupRepository.findAll()).thenReturn(Arrays.asList(entry1, entry2, entry3));
 
         RetentionEntry retentionEntry = new RetentionEntry("volumeId 1", 0, 0, 1);
-        when(retentionRepository.findByInstanceId(null)).thenReturn(Arrays.asList(retentionEntry));
+        when(retentionRepository.findAll()).thenReturn(Arrays.asList(retentionEntry));
 
         retentionService.apply();
 
@@ -219,19 +219,19 @@ public class RetentionServiceImplTest {
 
     @Test
     public void removeEmptyRetentionsTest() {
-        BackupEntry entry1 = new BackupEntry("volumeId 1", "fileName1", new DateTime().minusDays(2).getMillis() + "", "1", COMPLETED, "", "", "", "", "");
-        BackupEntry entry2 = new BackupEntry("volumeId 2", "fileName2", new DateTime().minusDays(23).getMillis() + "", "1", COMPLETED, "", "", "", "", "");
-        BackupEntry entry3 = new BackupEntry("volumeId 3", "fileName3", new DateTime().plusDays(2).getMillis() + "", "1", COMPLETED, "", "", "", "", "");
-        when(backupRepository.findAll(null)).thenReturn(Arrays.asList(entry1, entry2, entry3));
+        BackupEntry entry1 = new BackupEntry("volumeId 1", "fileName1", new DateTime().minusDays(2).getMillis() + "", "1", COMPLETED, "", "", "", "");
+        BackupEntry entry2 = new BackupEntry("volumeId 2", "fileName2", new DateTime().minusDays(23).getMillis() + "", "1", COMPLETED, "", "", "", "");
+        BackupEntry entry3 = new BackupEntry("volumeId 3", "fileName3", new DateTime().plusDays(2).getMillis() + "", "1", COMPLETED, "", "", "", "");
+        when(backupRepository.findAll()).thenReturn(Arrays.asList(entry1, entry2, entry3));
 
         RetentionEntry retentionEntry = new RetentionEntry("volumeId 1", 2, 0, 0);
         RetentionEntry retentionEntry2 = new RetentionEntry("volumeId 2", 0, 0, 0);
         RetentionEntry retentionEntry4 = new RetentionEntry("volumeId 4", 0, 0, 0);
-        when(retentionRepository.findByInstanceId(null)).thenReturn(Arrays.asList(retentionEntry, retentionEntry2, retentionEntry4));
+        when(retentionRepository.findAll()).thenReturn(Arrays.asList(retentionEntry, retentionEntry2, retentionEntry4));
 
         retentionService.apply();
 
-        verify(retentionRepository).delete(RetentionEntry.getId("volumeId 2", null));
-        verify(retentionRepository).delete(RetentionEntry.getId("volumeId 4", null));
+        verify(retentionRepository).delete("volumeId 2");
+        verify(retentionRepository).delete("volumeId 4");
     }
 }
